@@ -2,7 +2,7 @@ package org.exthmui.updater.controller;
 
 import android.content.Context;
 import android.content.Intent;
-import android.support.v4.content.LocalBroadcastManager;
+import androidx.localbroadcastmanager.content.LocalBroadcastManager;
 import android.util.Log;
 
 import org.exthmui.updater.model.Notice;
